+++ spec/spec_helper.rb
@@ -50,8 +50,8 @@ RSpec.configure do |config|
     if @fake_logger.crashes?
       crashes = @fake_logger.crashes.map do |args, call_stack|
         msg, ex = *args
-        "\n** Crash: #{msg.inspect}(#{ex.inspect})\n  Backtrace:\n    (crash) #{call_stack * "\n    (crash) " }"\
-          "\n** Crash: \"Actor crashed!\"(#{ex.inspect})\n  Backtrace:\n    (crash) #{call_stack * "\n    (crash) " }"\
+        "\n** Crash: #{msg.inspect}(#{ex.inspect})\n  Backtrace:\n    (crash) #{call_stack * "\n    (crash) "}"\
+          "\n** Crash: \"Actor crashed!\"(#{ex.inspect})\n  Backtrace:\n    (crash) #{call_stack * "\n    (crash) "}"\
           "\n  Exception Backtrace (#{ex.inspect}):\n    (ex) #{ex.backtrace * "\n    (ex) "}"
       end.join("\n")